FixLinuxWarring

+++ HBase/HLogic/HFunction.cpp
@@ -68,7 +68,7 @@ HObject* HFunction::addArgs(HArgs args)
 
 HObject* HFunction::hexec(HArgs args)
 {
-	CheckArgs(argsname.size());
+	CheckArgs((int)argsname.size());
 	for (int i = 0; i < argsname.size(); i++)
 		def->importclass(*argsname[i], args[i]);
 	for (int i = 0; i < commands.length(); i++)
